refactor myx-cli into module-based command handlers
