+++ crates/myx-cli/src/commands/mod.rs
@@ -1,0 +1,9 @@
+pub mod add;
+pub mod build;
+pub mod init;
+pub mod inspect;
+
+pub use add::command_add;
+pub use build::command_build;
+pub use init::command_init;
+pub use inspect::command_inspect;
